+++ src/parameters/quantity.rs
@@ -25,11 +25,11 @@ use std::ops::{Div, Mul};
 #[derive(Deserialize, Serialize, Clone, Copy, Default, Debug)]
 pub struct Units {
     /// Exponent for force.
-    pub f: f32,
+    pub f: f64,
     /// Exponent for length.
-    pub l: f32,
+    pub l: f64,
     /// Exponent for time.
-    pub t: f32,
+    pub t: f64,
 }
 
 impl PartialEq for Units {
@@ -65,7 +65,7 @@ impl Units {
     ///     l: 0.0 * n,
     /// };
     /// ```
-    fn pow(&self, exp: f32) -> Units {
+    fn pow(&self, exp: f64) -> Units {
         Units {
             f: exp * self.f,
             l: exp * self.l,
@@ -186,7 +186,7 @@ impl Div for Units {
 pub trait Quantity {
     /// Return a quantity that whose number part is a multiple
     /// of this quantity.
-    fn mul_number(&self, multiple: f32) -> Self;
+    fn mul_number(&self, multiple: f64) -> Self;
 
     /// Return a quantity that is `10^-3` times the original.
     fn kilo(&self) -> Self;
@@ -197,8 +197,8 @@ pub trait Quantity {
     /// Return a quantity that is `10^-9` times the original.
     fn nano(&self) -> Self;
 
-    /// Return the number (`f32`) part of the quantity (not the units).
-    fn number(&self) -> f32;
+    /// Return the number (`f64`) part of the quantity (not the units).
+    fn number(&self) -> f64;
 
     /// Return the `Units` of the `Quantity`
     fn units(&self) -> Units;
@@ -208,7 +208,7 @@ pub trait Quantity {
 
     /// Return a quantity that is the `exp`th power of
     /// this quantity.
-    fn pow(&self, exp: f32) -> General;
+    fn pow(&self, exp: f64) -> General;
 }
 
 /// A general quantity.
@@ -217,7 +217,7 @@ pub trait Quantity {
 )]
 pub struct General {
     /// Numerical value of this quantity.
-    n: f32,
+    n: f64,
     /// Units of this quantity.
     u: Units,
 }
@@ -259,7 +259,7 @@ impl General {
 }
 
 impl Quantity for General {
-    fn mul_number(&self, multiplier: f32) -> Self {
+    fn mul_number(&self, multiplier: f64) -> Self {
         General {
             n: self.n * multiplier,
             u: self.u,
@@ -278,7 +278,7 @@ impl Quantity for General {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.n
     }
 
@@ -290,7 +290,7 @@ impl Quantity for General {
         *self
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         General {
             n: self.n.powf(exp),
             u: self.u.pow(exp),
@@ -329,10 +329,10 @@ impl Display for General {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Force(pub f32);
+pub struct Force(pub f64);
 
 impl Quantity for Force {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Force(self.0 * other)
     }
 
@@ -348,7 +348,7 @@ impl Quantity for Force {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -363,7 +363,7 @@ impl Quantity for Force {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -371,10 +371,10 @@ impl Quantity for Force {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Length(pub f32);
+pub struct Length(pub f64);
 
 impl Quantity for Length {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Length(self.0 * other)
     }
 
@@ -390,7 +390,7 @@ impl Quantity for Length {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -405,7 +405,7 @@ impl Quantity for Length {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -413,10 +413,10 @@ impl Quantity for Length {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Time(pub f32);
+pub struct Time(pub f64);
 
 impl Quantity for Time {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Time(self.0 * other)
     }
 
@@ -432,7 +432,7 @@ impl Quantity for Time {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -447,7 +447,7 @@ impl Quantity for Time {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -455,10 +455,10 @@ impl Quantity for Time {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Tinv(pub f32);
+pub struct Tinv(pub f64);
 
 impl Quantity for Tinv {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Tinv(self.0 * other)
     }
 
@@ -474,7 +474,7 @@ impl Quantity for Tinv {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -489,7 +489,7 @@ impl Quantity for Tinv {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -497,10 +497,10 @@ impl Quantity for Tinv {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Diffusion(pub f32);
+pub struct Diffusion(pub f64);
 
 impl Quantity for Diffusion {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Diffusion(self.0 * other)
     }
 
@@ -516,7 +516,7 @@ impl Quantity for Diffusion {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -531,7 +531,7 @@ impl Quantity for Diffusion {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -539,10 +539,10 @@ impl Quantity for Diffusion {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Stress(pub f32);
+pub struct Stress(pub f64);
 
 impl Quantity for Stress {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Stress(self.0 * other)
     }
 
@@ -558,7 +558,7 @@ impl Quantity for Stress {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -573,7 +573,7 @@ impl Quantity for Stress {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
@@ -581,10 +581,10 @@ impl Quantity for Stress {
 #[derive(
     Deserialize, Serialize, Clone, Copy, PartialEq, Default, Debug,
 )]
-pub struct Viscosity(pub f32);
+pub struct Viscosity(pub f64);
 
 impl Quantity for Viscosity {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Viscosity(self.0 * other)
     }
 
@@ -600,7 +600,7 @@ impl Quantity for Viscosity {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -615,16 +615,16 @@ impl Quantity for Viscosity {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
 
 #[derive(Deserialize, Clone, Copy, PartialEq, Default, Debug)]
-pub struct Unitless(pub f32);
+pub struct Unitless(pub f64);
 
 impl Quantity for Unitless {
-    fn mul_number(&self, other: f32) -> Self {
+    fn mul_number(&self, other: f64) -> Self {
         Unitless(self.0 * other)
     }
 
@@ -640,7 +640,7 @@ impl Quantity for Unitless {
         self.mul_number(1e-9)
     }
 
-    fn number(&self) -> f32 {
+    fn number(&self) -> f64 {
         self.0
     }
 
@@ -655,7 +655,7 @@ impl Quantity for Unitless {
         }
     }
 
-    fn pow(&self, exp: f32) -> General {
+    fn pow(&self, exp: f64) -> General {
         self.g().pow(exp)
     }
 }
